backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
neo4j==5.13.0
spacy==3.7.2
jieba==0.42.1
pydantic==2.5.0
sqlalchemy==2.0.23
pytesseract==0.3.10
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.1.0
pandas==2.1.4
numpy==1.24.3
aiofiles==23.2.0
jinja2==3.1.2
requests==2.31.0
email-validator==2.1.0